No longer overrides global setting for jquery.metadata plugin. Closes gh-5
Allows mbmenu to be used in applications that use one of the other available storage options of jquery.metadata.
Menu settings can be loaded via the jquery.metadata plugin, regardless of how the metadata is stored in the HTML.
There is no need to force the metadata to be loaded from the class attribute, since that is an implementation detail of the metadata plugin.